Revert "(improvement) Move layouts from localStorage to ws"
