+++ src/components/HFUI/HFUI.container.js
@@ -33,12 +33,6 @@ const mapDispatchToProps = dispatch => ({
       mode,
     ]))
   },
-  getLayouts: (authToken) => {
-    dispatch(WSActions.send([
-      'layouts.get',
-      authToken,
-    ]))
-  },
   onUnload: (authToken, mode) => {
     dispatch(WSActions.onUnload(authToken, mode))
   },
