--- src/components/LoadingPanel/LoadingPanel.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import Panel from '../../ui/Panel'
-
-import './style.css'
-
-const LoadingPanel = ({ dark }) => (
-  <Panel
-    dark={dark}
-    darkHeader={dark}
-    className='loading-panel'
-  />
-)
-
-export default LoadingPanel
