--- src/components/LoadingPanel/index.js
@@ -1,3 +1,0 @@
-import LoadingPanel from './LoadingPanel'
-
-export default LoadingPanel
