--- src/components/LoadingPanel/style.scss
@@ -1,5 +1,0 @@
-.loading-panel {
-  .hfui-panel__header {
-    display: none;
-  }
-}
