+++ src/redux/constants/ui.js
@@ -7,7 +7,6 @@ module.exports = {
   CREATE_LAYOUT: 'UI_CREATE_LAYOUT',
   DELETE_LAYOUT: 'UI_DELETE_LAYOUT',
 
-  SET_LAYOUTS: 'UI_SET_LAYOUTS',
   SET_LAYOUT_ID: 'UI_SET_LAYOUT_ID',
   ADD_COMPONENT: 'UI_ADD_COMPONENT',
   REMOVE_COMPONENT: 'UI_REMOVE_COMPONENT',
