+++ src/redux/constants/ws.js
@@ -31,7 +31,6 @@ export default {
   DATA_ALGO_ORDER: 'WS_DATA_ALGO_ORDER',
   DATA_ALGO_ORDERS: 'WS_DATA_ALGO_ORDERS',
   DATA_NOTIFICATION: 'WS_DATA_NOTIFICATION',
-  DATA_LAYOUTS: 'WS_DATA_LAYOUTS',
   CLEAR_ALGO_ORDERS: 'WS_CLEAR_ALGO_ORDERS',
 
   PURGE_DATA_BACKTEST: 'WS_PURGE_DATA_BACKTEST',
@@ -50,7 +49,6 @@ export default {
   SET_BACKTEST_OPTIONS: 'SET_BACKTEST_OPTIONS',
   SET_BACKTEST_LOADING: 'WS_SET_BACKTEST_LOADING',
   UPDATE_FAVORITE_PAIRS: 'UPDATE_FAVORITE_PAIRS',
-  UPDATE_LAYOUTS: 'UPDATE_LAYOUTS',
 
   BUFFER_DATA_FROM_EXCHANGE: 'WS_BUFFER_DATA_FROM_EXCHANGE',
