+++ src/redux/reducers/ui/default_layout_market_data.js
@@ -1,0 +1,19 @@
+import { marketData } from '../../../constants/routes'
+
+export default {
+  routePath: marketData.path,
+  canDelete: false,
+  isDefault: true,
+  savedAt: 0,
+  layout: [{
+    w: 53, h: 10, minW: 30, minH: 8, x: 0, y: 0, i: '1563115202169', moved: false, static: false, c: 'CHART',
+  }, {
+    w: 24, h: 20, minW: 21, minH: 5, x: 76, y: 0, i: '1563119067081', moved: false, static: false, c: 'ORDER_BOOK',
+  }, {
+    w: 23, h: 10, minW: 18, minH: 4, x: 53, y: 0, i: '1563119098026', moved: false, static: false, c: 'TRADES_TABLE',
+  }, {
+    w: 53, h: 10, minW: 30, minH: 8, x: 0, y: 10, i: '1565595676626', moved: false, static: false, c: 'CHART',
+  }, {
+    w: 23, h: 10, minW: 18, minH: 4, x: 53, y: 10, i: '1565595694304', moved: false, static: false, c: 'TRADES_TABLE',
+  }],
+}
