+++ src/redux/reducers/ws/index.js
@@ -11,7 +11,6 @@ import orderHistory from './order_history'
 import algoOrders from './algo_orders'
 import backtest from './backtest'
 import favoriteTradingPairs from './favorite_pairs'
-import layouts from './layouts'
 
 export default combineReducers({
   algoOrders,
@@ -25,5 +24,4 @@ export default combineReducers({
   auth,
   backtest,
   favoriteTradingPairs,
-  layouts,
 })
