--- src/redux/selectors/ui/get_is_ws_layouts_set.js
@@ -1,8 +1,0 @@
-import _get from 'lodash/get'
-import { REDUCER_PATHS } from '../../config'
-
-const path = REDUCER_PATHS.UI
-
-export default (state) => {
-  return _get(state, `${path}.isWsLayoutsSet`)
-}
