+++ src/redux/selectors/ui/index.js
@@ -5,10 +5,10 @@ import getActiveMarket from './get_active_market'
 import getLayouts from './get_layouts'
 import getLayout from './get_layout'
 import getLayoutID from './get_layout_id'
-import getIsWsLayoutsSet from './get_is_ws_layouts_set'
 import getRemoteVersion from './get_remote_version'
 import getIsTradingModeModalVisible from './get_is_trading_mode_modal_visible'
 import getIsInternetConnection from './get_is_bad_internet_connection'
+import getIsAOPausedModalVisible from './get_is_ao_paused'
 import getOldFormatModalState from './get_old_modal_format_state'
 import getIsRefillBalanceModalVisible from './get_is_refill_balance_modal_visible'
 import getIsPaperTrading from './get_is_paper_trading'
@@ -20,7 +20,6 @@ import getStrategyId from './get_strategy_id'
 import getMarketComponents from './get_market_components'
 import getCurrentUnsavedLayout from './get_current_unsaved_layout'
 import getTickersVolumeUnit from './get_tickers_volume_unit'
-import getIsAOPausedModalVisible from './get_is_ao_paused'
 import getIsCcyInfoModalVisible from './get_is_ccy_info_modal_visible'
 
 export * from './get_settings'
@@ -34,7 +33,6 @@ export {
   getLayouts,
   getLayout,
   getLayoutID,
-  getIsWsLayoutsSet,
   getIsTradingModeModalVisible,
   getIsInternetConnection,
   getIsRefillBalanceModalVisible,
