--- src/redux/selectors/ws/get_layouts_loaded.js
@@ -1,8 +1,0 @@
-import _get from 'lodash/get'
-import { REDUCER_PATHS } from '../../config'
-
-const path = REDUCER_PATHS.WS
-
-export default (state) => {
-  return _get(state, `${path}.layouts.isLoaded`, false)
-}
