+++ src/redux/selectors/ws/index.js
@@ -21,8 +21,6 @@ import getFilteredAlgoOrdersCount from './get_filtered_algo_orders_count'
 import getAlgoOrders from './get_algo_orders'
 import getNotifications from './get_notifications'
 import getAllSyncRanges from './get_all_sync_ranges'
-import getLayouts from './get_layouts'
-import getLayoutsLoaded from './get_layouts_loaded'
 import getOrderHistory from './get_order_history'
 
 import getBacktestState from './get_backtest_state'
@@ -75,8 +73,6 @@ export {
   getFilteredAlgoOrdersCount,
   getFavoritePairs,
   getFavoritePairsObject,
-  getLayouts,
-  getLayoutsLoaded,
 
   getNotifications,
   getAllSyncRanges,
